Added Basic Movement and Animator components
Player keyboard input is handled in the "PlayerInputController" class which uses the KeyCodes set in the "PlayerControls" class which will be used in the future to allow for custom control layouts for the user. "PlayerInputController" then sets the neccessary values in the "PlayerMovementController" which handles the movement of the "Player" class. Also included in the changes is a "PlayerPropertiesController" which will be used in the future to modify the Players stats.